requirements.txt
pyTelegramBotAPI==4.22.1
Flask==3.0.3
python-dotenv==1.0.1
pymongo==4.8.0
dnspython==2.6.1
